+++ integrationtest/dh/user_test.go
@@ -36,7 +36,7 @@ func TestUserCreationAndObtaining(t *testing.T) {
 	}
 
 	is.True(currentUser != nil)
-	is.Equal(currentUser.Id, int64(*userId))
+	is.Equal(currentUser.Id, *userId)
 
 	list, err := client.ListUsers(&dh.ListParams{
 		UserStatus: 0,
