+++ notification.go
@@ -6,10 +6,10 @@ type notificationResponse struct {
 }
 
 type Notification struct {
-	Id           int64                  `json:"id"`
+	Id           int                    `json:"id"`
 	Notification string                 `json:"notification"`
 	Timestamp    ISO8601Time            `json:"timestamp"`
 	DeviceId     string                 `json:"deviceId"`
-	NetworkId    int64                  `json:"networkId"`
+	NetworkId    int                    `json:"networkId"`
 	Parameters   map[string]interface{} `json:"parameters"`
 }
